fix: إصلاح خطأ تعيين النوع في smart_sync_manager (السطر 286)

+++ mobile/lib/services/smart_sync_manager.dart
@@ -283,7 +283,7 @@ class SmartSyncManager {
       deviceId = deviceIdValue is String ? deviceIdValue : deviceIdValue?.toString();
       final recordsCountValue = newBackup.appProperties['records_count'];
       if (recordsCountValue is int) {
-        recordsCount = recordsCountValue;
+        recordsCount = recordsCountValue as int;
       } else if (recordsCountValue is String) {
         recordsCount = int.tryParse(recordsCountValue);
       }
